Splitted code to different files
